# openapi/baseline.json
{
  "components": {
    "schemas": {
...
      "ClusterEdgeDto": {
        "properties": {
          "conversationCount": {
            "format": "int64",
            "type": "integer"
          },
          "dominantProtocol": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "totalBytes": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ClusterGraphResponse": {
        "properties": {
          "clusters": {
            "items": {
              "$ref": "#/components/schemas/ClusterNodeDto"
            },
            "type": "array"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/ClusterEdgeDto"
            },
            "type": "array"
          },
          "groupType": {
            "type": "string"
          },
          "hiddenClusters": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "ClusterNodeDto": {
        "properties": {
          "conversationCount": {
            "format": "int64",
            "type": "integer"
          },
          "dominantProtocols": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "geoSource": {
            "type": "string"
          },
          "groupType": {
            "type": "string"
          },
          "hostCount": {
            "format": "int64",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "ipBytes": {
            "additionalProperties": {
              "format": "int64",
              "type": "integer"
            },
            "type": "object"
          },
          "ipConversations": {
            "additionalProperties": {
              "format": "int64",
              "type": "integer"
            },
            "type": "object"
          },
          "ipPeers": {
            "additionalProperties": {
              "format": "int64",
              "type": "integer"
            },
            "type": "object"
          },
          "ipRisks": {
            "additionalProperties": {
              "format": "int64",
              "type": "integer"
            },
            "type": "object"
          },
          "label": {
            "type": "string"
          },
          "lat": {
            "format": "double",
            "type": "number"
          },
          "lon": {
            "format": "double",
            "type": "number"
          },
          "riskCount": {
            "format": "int64",
            "type": "integer"
          },
          "sampleIps": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "topRiskTypes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "totalBytes": {
            "format": "int64",
            "type": "integer"
          },
          "totalPackets": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "DnsQueryEntryDto": {
        "properties": {
          "frame": {
            "format": "int64",
            "type": "integer"
          },
          "queryCount": {
            "format": "int32",
            "type": "integer"
          },
          "queryName": {
            "type": "string"
          },
          "queryType": {
            "type": "string"
          },
          "resolvable": {
            "type": "boolean"
          },
          "resolvedIps": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "responseCode": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "DnsQueryLogResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DnsQueryEntryDto"
            },
            "type": "array"
          },
          "failedCount": {
            "format": "int64",
            "type": "integer"
          },
          "hostname": {
            "type": "string"
          },
          "nxdomainRatio": {
            "format": "double",
            "type": "number"
          },
          "resolvedCount": {
            "format": "int64",
            "type": "integer"
          },
          "serverIp": {
            "type": "string"
          },
          "suspicious": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "HostSummaryDto": {
        "properties": {
          "conversationCount": {
            "format": "int64",
            "type": "integer"
          },
          "country": {
            "type": "string"
          },
          "deviceType": {
            "type": "string"
          },
          "geoSource": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "hostnameSource": {
            "type": "string"
          },
          "ip": {
            "type": "string"
          },
          "org": {
            "type": "string"
          },
          "packetCount": {
            "format": "int64",
            "type": "integer"
          },
          "riskCount": {
            "format": "int64",
            "type": "integer"
          },
          "role": {
            "type": "string"
          },
          "totalBytes": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
      "HttpEndpointDto": {
        "properties": {
          "clientErrorCount": {
            "format": "int32",
            "type": "integer"
          },
          "contentType": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "requestCount": {
            "format": "int32",
            "type": "integer"
          },
          "requestFrame": {
            "format": "int64",
            "type": "integer"
          },
          "responseFrame": {
            "format": "int64",
            "type": "integer"
          },
          "serverErrorCount": {
            "format": "int32",
            "type": "integer"
          },
          "successCount": {
            "format": "int32",
            "type": "integer"
          },
          "topStatus": {
            "format": "int32",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "PacketLocationResponse": {
        "properties": {
          "conversationId": {
            "format": "uuid",
            "type": "string"
          },
          "packetNumber": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "ServiceServerSummaryDto": {
        "properties": {
          "anomalyRatio": {
            "format": "double",
            "type": "number"
          },
          "failedCount": {
            "format": "int64",
            "type": "integer"
          },
          "hostname": {
            "type": "string"
          },
          "okCount": {
            "format": "int64",
            "type": "integer"
          },
          "role": {
            "type": "string"
          },
          "serverIp": {
            "type": "string"
          },
          "suspicious": {
            "type": "boolean"
          },
          "totalRequests": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
      "TlsInfo": {
        "properties": {
          "issuer": {
            "type": "string"
          },
          "ja3s": {
            "type": "string"
          },
          "notAfter": {
            "format": "date-time",
            "type": "string"
          },
          "notBefore": {
            "format": "date-time",
            "type": "string"
          },
          "sniNames": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "subject": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "TopHostsResponse": {
        "properties": {
          "hosts": {
            "items": {
              "$ref": "#/components/schemas/HostSummaryDto"
            },
            "type": "array"
          }
        },
        "type": "object"
      },
...
      "WebServerDetailResponse": {
        "properties": {
          "api": {
            "type": "boolean"
          },
          "clientErrorCount": {
            "format": "int64",
            "type": "integer"
          },
          "clientErrorRatio": {
            "format": "double",
            "type": "number"
          },
          "contentTypes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "endpoints": {
            "items": {
              "$ref": "#/components/schemas/HttpEndpointDto"
            },
            "type": "array"
          },
          "hostname": {
            "type": "string"
          },
          "serverErrorCount": {
            "format": "int64",
            "type": "integer"
          },
          "serverIp": {
            "type": "string"
          },
          "serverSoftware": {
            "type": "string"
          },
          "successCount": {
            "format": "int64",
            "type": "integer"
          },
          "suspicious": {
            "type": "boolean"
          },
          "tls": {
            "$ref": "#/components/schemas/TlsInfo"
          },
          "totalRequests": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      }
    }
  },
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/intelligence/{fileId}/clusters": {
      "get": {
        "description": "Returns network hosts grouped into clusters by ASN, country, subnet, or device type. Supports the same conversation filters as the conversations endpoint to pre-filter traffic before clustering.",
        "operationId": "getClusters",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "groupBy",
            "required": false,
            "schema": {
              "default": "asn",
              "type": "string"
            }
          },
          {
            "description": "Filter by IP address or hostname",
            "in": "query",
            "name": "ip",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by port number (src or dst)",
            "in": "query",
            "name": "port",
            "required": false,
            "schema": {
              "format": "int32",
              "maximum": 65535,
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "description": "Comma-separated L4 protocols",
            "in": "query",
            "name": "protocols",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated L7 protocols",
            "in": "query",
            "name": "l7Protocols",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated application names",
            "in": "query",
            "name": "apps",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated categories",
            "in": "query",
            "name": "categories",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Only conversations with flow risks",
            "in": "query",
            "name": "hasRisks",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "description": "Comma-separated detected file types",
            "in": "query",
            "name": "fileTypes",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated nDPI risk types",
            "in": "query",
            "name": "riskTypes",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated custom signature rule names",
            "in": "query",
            "name": "customSignatures",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Filter by payload content (ASCII or hex)",
            "in": "query",
            "name": "payloadContains",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated device types",
            "in": "query",
            "name": "deviceTypes",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated ISO 3166-1 alpha-2 country codes",
            "in": "query",
            "name": "countries",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Comma-separated network label names (e.g. 'Office,DMZ')",
            "in": "query",
            "name": "networkLabels",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterGraphResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get clustered network topology",
        "tags": [
          "Network Intelligence"
        ]
      }
    },
    "/api/v1/intelligence/{fileId}/dns-servers": {
      "get": {
        "description": "Returns every host that answered DNS queries in the capture, with resolved vs. failed counts and an NXDOMAIN-based suspicious flag (possible DNS tunnelling / domain-generation algorithm).",
        "operationId": "getDnsServers",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceServerSummaryDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List DNS servers and their resolution health",
        "tags": [
          "Network Intelligence"
        ]
      }
    },
    "/api/v1/intelligence/{fileId}/dns/{serverIp}": {
      "get": {
        "description": "Returns the per-domain query log (hostname queried, response code, resolved IPs, query count, resolvable) for the given DNS server, plus summary counts and the suspicious verdict.",
        "operationId": "getDnsQueryLog",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "serverIp",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/DnsQueryLogResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the DNS query log for one DNS server",
        "tags": [
          "Network Intelligence"
        ]
      }
    },
    "/api/v1/intelligence/{fileId}/packet-location/{packetNumber}": {
      "get": {
        "description": "Returns the conversation that contains the given packet (frame number) so the UI can open and highlight it.",
        "operationId": "locatePacket",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "packetNumber",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PacketLocationResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Locate a packet by frame number",
        "tags": [
          "Network Intelligence"
        ]
      }
    },
    "/api/v1/intelligence/{fileId}/top-hosts": {
      "get": {
        "description": "Returns the top N hosts ranked by bytes, packets, conversations, or risk count.",
        "operationId": "getTopHosts",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "required": false,
            "schema": {
              "default": "bytes",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "format": "int32",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TopHostsResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get top hosts by traffic volume",
        "tags": [
          "Network Intelligence"
        ]
      }
    },
    "/api/v1/intelligence/{fileId}/web-servers": {
      "get": {
        "description": "Returns every host classified as a web/API server (includes HTTPS-only hosts), with success vs. error response counts and a 4xx-based suspicious flag (possible endpoint enumeration / scanning).",
        "operationId": "getWebServers",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceServerSummaryDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "List web/API servers and their HTTP health",
        "tags": [
          "Network Intelligence"
        ]
      }
    },
    "/api/v1/intelligence/{fileId}/web/{serverIp}": {
      "get": {
        "description": "Returns the per-endpoint log (method, path, status-class counts, content type) plus server software, content types and TLS metadata (cleartext HTTP only for endpoints; TLS detail for HTTPS).",
        "operationId": "getWebServerDetail",
        "parameters": [
          {
            "in": "path",
            "name": "fileId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "serverIp",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/WebServerDetailResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get the HTTP endpoint log + detail for one web/API server",
        "tags": [
          "Network Intelligence"
        ]
      }
    },
...
  },
...
}